Adding configurator.
